requirements.txt
requests==2.31.0
python-dotenv==1.0.0
authlib==1.2.1
flask==2.3.3
sqlalchemy==2.0.36
apscheduler==3.10.4